templates/rag-03-llamaindex/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
python-multipart==0.0.9
pydantic-settings>=2.5.2
python-dotenv==1.0.1
llama-index-core==0.12.5
llama-index-llms-openai==0.3.18
llama-index-embeddings-openai==0.3.1
llama-index-vector-stores-chroma==0.4.1
llama-index-readers-file==0.4.4
llama-index-readers-web==0.3.2
chromadb==0.5.15
pymupdf==1.25.0
mcp[cli]>=1.6.0
httpx>=0.27.0
